chore: Add footer and header per page basis
The pages that require either footer or and header will be applied
individually.
This is due to the issues with footer stiking and sometimes hindering
view.
The header may not be needed in other places like signin and signout.
The dashboard, login and logout pages have received a redesign.
The new style may not persist much longer though as new the site is
constantly evolving.

Signed-off-by: Gekko Wrld <108436933+gekkowrld@users.noreply.github.com>