Display product review and style (#15)
* chore: Display product review

* Restyle Display product review (#16)

* Restyled by clang-format

* Restyled by jq

* Restyled by prettier

* Restyled by prettier-json

* chore: Add review page

* Restyle Display product review (#17)

* Restyled by clang-format

* Restyled by jq

* Restyled by prettier

* Restyled by prettier-json

* Update App.js

* fix: Add aria-label description

* chore: Add footer and header per page basis

The pages that require either footer or and header will be applied
individually.
This is due to the issues with footer stiking and sometimes hindering
view.
The header may not be needed in other places like signin and signout.
The dashboard, login and logout pages have received a redesign.
The new style may not persist much longer though as new the site is
constantly evolving.


Signed-off-by: Gekko Wrld <108436933+gekkowrld@users.noreply.github.com>
Co-authored-by: restyled-io[bot] <32688539+restyled-io[bot]@users.noreply.github.com>
Co-authored-by: Restyled.io <commits@restyled.io>