fix: Ensure no duplicate data is entered
I have ensured that no duplicate data can be entered in "reviews"
collection.
The only problem now is figuring out if the there is an entry already.
This is simple in theory, but it has proved challenging for now.
Cart ui part 1 is done for now.
The other parts like the actual functionality have not yet been added.
The payment method is also in consideration as of now.

Signed-off-by: Gekko Wrld <108436933+gekkowrld@users.noreply.github.com>