Mod (#19)
* fix: Ensure no duplicate data is entered

I have ensured that no duplicate data can be entered in "reviews"
collection.
The only problem now is figuring out if the there is an entry already.
This is simple in theory, but it has proved challenging for now.
Cart ui part 1 is done for now.
The other parts like the actual functionality have not yet been added.
The payment method is also in consideration as of now.

* docs: Add school spec in the readme file

* chore: Add checkout option

* fix: Use absolute path instead of relative to avoid potential 404's

* chore: Add save functionality